create useCustomResolver hook
